Add files via upload
Heart disease, also known as cardiovascular disease, is one of the most serious illnesses in both India and the rest of the globe. According to estimates, cardiac illnesses account for 28.1% of fatalities. More than 17.6 million fatalities, or a large portion of all deaths in 2016, were caused by it in 2016. Therefore, a system that can predict with exact precision and of such disease. Numerous academics do extensive research utilizing a variety of machine learning algorithms to predict heart illness using different datasets that contain numerous factors that lead to heart attacks